requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.23.5
joblib==1.3.1
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.6.5
shap==0.41.0
xgboost==2.1.3
tldextract
dnspython
requests
beautifulsoup4
python-whois
python-dotenv
python-Levenshtein
word_with_nlp
sklearn-compat
